Issue-8: More test cases.  Convert to parameterized test.
